+++ skbio/diversity/alpha/_base.py
@@ -14,37 +14,23 @@ from scipy.optimize import fmin_powell, minimize_scalar
 
 from skbio.stats import subsample_counts
 from skbio.util._decorator import experimental
-
-
-def _validate(counts, suppress_cast=False):
-    """Validate and convert input to an acceptable counts vector type.
-
-    Note: may not always return a copy of `counts`!
-
-    """
-    counts = np.asarray(counts)
-
-    if not suppress_cast:
-        counts = counts.astype(int, casting='safe', copy=False)
-
-    if counts.ndim != 1:
-        raise ValueError("Only 1-D vectors are supported.")
-    elif (counts < 0).any():
-        raise ValueError("Counts vector cannot contain negative values.")
-
-    return counts
+from skbio.diversity._base import _validate_counts_vector
 
 
 @experimental(as_of="0.4.0")
 def berger_parker_d(counts):
-    """Calculate Berger-Parker dominance.
+    r"""Calculate Berger-Parker dominance.
 
     Berger-Parker dominance is defined as the fraction of the sample that
-    belongs to the most abundant OTUs:
+    belongs to the most abundant OTU:
 
     .. math::
 
-       d = \\frac{N_{max}}{N}
+       d = \frac{N_{max}}{N}
+
+    where :math:`N_{max}` is defined as the number of individuals in the most
+    abundant OTU (or any of the most abundant OTUs in the case of ties), and
+    :math:`N` is defined as the total number of individuals in the sample.
 
     Parameters
     ----------
@@ -67,18 +53,23 @@ def berger_parker_d(counts):
     .. [2] http://www.pisces-conservation.com/sdrhelp/index.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return counts.max() / counts.sum()
 
 
 @experimental(as_of="0.4.0")
 def brillouin_d(counts):
-    """Calculate Brillouin index of alpha diversity, which is defined as:
+    r"""Calculate Brillouin index of alpha diversity.
+
+    This is calculated as follows:
 
     .. math::
 
-       HB = \\frac{\\ln N!-\\sum^5_{i=1}{\\ln n_i!}}{N}
+       HB = \frac{\ln N!-\sum^s_{i=1}{\ln n_i!}}{N}
 
+    where :math:`N` is defined as the total number of individuals in the
+    sample, :math:`s` is the number of OTUs, and :math:`n_i` is defined as the
+    number of individuals in the :math:`i^{\text{th}}` OTU.
 
     Parameters
     ----------
@@ -100,7 +91,7 @@ def brillouin_d(counts):
     .. [1] http://www.pisces-conservation.com/sdrhelp/index.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     nz = counts[counts.nonzero()]
     n = nz.sum()
     return (gammaln(n + 1) - gammaln(nz + 1).sum()) / n
@@ -108,13 +99,13 @@ def brillouin_d(counts):
 
 @experimental(as_of="0.4.0")
 def dominance(counts):
-    """Calculate dominance.
+    r"""Calculate dominance.
 
     Dominance is defined as
 
     .. math::
 
-       \\sum{p_i^2}
+       \sum{p_i^2}
 
     where :math:`p_i` is the proportion of the entire community that OTU
     :math:`i` represents.
@@ -145,7 +136,7 @@ def dominance(counts):
     .. [1] http://folk.uio.no/ohammer/past/diversity.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     freqs = counts / counts.sum()
     return (freqs * freqs).sum()
 
@@ -165,15 +156,22 @@ def doubles(counts):
         Doubleton count.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return (counts == 2).sum()
 
 
 @experimental(as_of="0.4.0")
 def enspie(counts):
-    """Calculate ENS_pie alpha diversity measure.
+    r"""Calculate ENS_pie alpha diversity measure.
 
-    ENS_pie is equivalent to ``1 / dominance``.
+    ENS_pie is equivalent to ``1 / dominance``:
+
+    .. math::
+
+       ENS_{pie} = \frac{1}{\sum_{i=1}^s{p_i^2}}
+
+    where :math:`s` is the number of OTUs and :math:`p_i` is the proportion of
+    the community represented by OTU :math:`i`.
 
     Parameters
     ----------
@@ -200,7 +198,7 @@ def enspie(counts):
        Ecology Letters, Volume 16, Issue Supplement s1, pgs 17-26.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return 1 / dominance(counts)
 
 
@@ -234,7 +232,7 @@ def equitability(counts, base=2):
     .. [1] http://www.pisces-conservation.com/sdrhelp/index.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     numerator = shannon(counts, base)
     denominator = np.log(observed_otus(counts)) / np.log(base)
     return numerator / denominator
@@ -242,13 +240,13 @@ def equitability(counts, base=2):
 
 @experimental(as_of="0.4.0")
 def esty_ci(counts):
-    """Calculate Esty's CI.
+    r"""Calculate Esty's CI.
 
     Esty's CI is defined as
 
     .. math::
 
-       F_1/N \\pm z\\sqrt{W}
+       F_1/N \pm z\sqrt{W}
 
     where :math:`F_1` is the number of singleton OTUs, :math:`N` is the total
     number of individuals (sum of abundances for all OTUs), and :math:`z` is a
@@ -259,7 +257,7 @@ def esty_ci(counts):
 
     .. math::
 
-       \\frac{F_1(N-F_1)+2NF_2}{N^3}
+       \frac{F_1(N-F_1)+2NF_2}{N^3}
 
     where :math:`F_2` is the number of doubleton OTUs.
 
@@ -284,7 +282,7 @@ def esty_ci(counts):
        estimator of the coverage of a random sample". Ann Statist 11: 905-912.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
 
     f1 = singles(counts)
     f2 = doubles(counts)
@@ -295,9 +293,65 @@ def esty_ci(counts):
     return f1 / n - z * np.sqrt(W), f1 / n + z * np.sqrt(W)
 
 
+@experimental(as_of="0.4.0-dev")
+def faith_pd(counts, otu_ids, tree):
+    """ Compute Faith's phylogenetic diversity metric (PD)
+
+    Parameters
+    ----------
+    counts : 1-D array_like, int
+        Vector of counts.
+    otu_ids: list, np.array
+        Vector of OTU ids corresponding to tip names in ``tree``. Must be the
+        same length as ``counts``.
+    tree: skbio.TreeNode
+        Tree relating the OTUs in otu_ids. The set of tip names in the tree can
+        be a superset of ``otu_ids``, but not a subset.
+
+    Returns
+    -------
+    float
+        The phylogenetic diversity (PD) of the samples.
+
+    Raises
+    ------
+    ValueError
+        If ``counts`` and ``otu_ids`` are not equal in length.
+    MissingNodeError
+        If an OTU id is provided that does not correspond to a tip in the
+        tree.
+
+    Notes
+    -----
+    Faith's phylogenetic diversity, often referred to as PD, was originally
+    described in [1]_.
+
+    References
+    ----------
+    .. [1] Faith, D. P. Conservation evaluation and phylogenetic diversity.
+       Biol. Conserv. (1992).
+
+    """
+    counts = _validate_counts_vector(counts)
+    observed_otus = {o: c for o, c in zip(otu_ids, counts) if c >= 1}
+    observed_nodes = tree.observed_node_counts(observed_otus)
+    result = sum(o.length for o in observed_nodes if o.length is not None)
+    return result
+
+
 @experimental(as_of="0.4.0")
 def fisher_alpha(counts):
-    """Calculate Fisher's alpha.
+    r"""Calculate Fisher's alpha, a metric of diversity.
+
+    Fisher's alpha is estimated by solving the following equation for
+    :math:`\alpha`:
+
+    .. math::
+
+       S=\alpha\ln(1+\frac{N}{\alpha})
+
+    where :math:`S` is the number of OTUs and :math:`N` is the
+    total number of individuals in the sample.
 
     Parameters
     ----------
@@ -325,7 +379,7 @@ def fisher_alpha(counts):
     .. [1] http://www.pisces-conservation.com/sdrhelp/index.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     n = counts.sum()
     s = observed_otus(counts)
 
@@ -349,13 +403,13 @@ def fisher_alpha(counts):
 
 @experimental(as_of="0.4.0")
 def goods_coverage(counts):
-    """Calculate Good's coverage of counts.
+    r"""Calculate Good's coverage of counts.
 
     Good's coverage estimator is defined as
 
     .. math::
 
-       1-\\frac{F_1}{N}
+       1-\frac{F_1}{N}
 
     where :math:`F_1` is the number of singleton OTUs and :math:`N` is the
     total number of individuals (sum of abundances for all OTUs).
@@ -371,7 +425,7 @@ def goods_coverage(counts):
         Good's coverage estimator.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     f1 = singles(counts)
     N = counts.sum()
     return 1 - (f1 / N)
@@ -379,7 +433,16 @@ def goods_coverage(counts):
 
 @experimental(as_of="0.4.0")
 def heip_e(counts):
-    """Calculate Heip's evenness measure.
+    r"""Calculate Heip's evenness measure.
+
+    Heip's evenness is defined as:
+
+    .. math::
+
+       \frac{(e^H-1)}{(S-1)}
+
+    where :math:`H` is the Shannon-Wiener entropy of counts (using logarithm
+    base :math:`e`) and :math:`S` is the number of OTUs in the sample.
 
     Parameters
     ----------
@@ -391,6 +454,10 @@ def heip_e(counts):
     double
         Heip's evenness measure.
 
+    See Also
+    --------
+    shannon
+
     Notes
     -----
     The implementation here is based on the description in [1]_.
@@ -401,7 +468,7 @@ def heip_e(counts):
        UK., 54, 555-557.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return ((np.exp(shannon(counts, base=np.e)) - 1) /
             (observed_otus(counts) - 1))
 
@@ -448,7 +515,7 @@ def kempton_taylor_q(counts, lower_quantile=0.25, upper_quantile=0.75):
     .. [2] http://www.pisces-conservation.com/sdrhelp/index.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     n = len(counts)
     lower = int(np.ceil(n * lower_quantile))
     upper = int(n * upper_quantile)
@@ -459,15 +526,16 @@ def kempton_taylor_q(counts, lower_quantile=0.25, upper_quantile=0.75):
 
 @experimental(as_of="0.4.0")
 def margalef(counts):
-    """Calculate Margalef's richness index, which is defined as:
+    r"""Calculate Margalef's richness index.
+
+    Margalef's D is defined as:
 
     .. math::
 
-       D = \\frac{(S - 1)}{\\ln N}
+       D = \frac{(S - 1)}{\ln N}
 
-    where :math:`S` is the species number and :math:`N` is the
-    total number of individuals (sum of abundances for all OTUs).
-
+    where :math:`S` is the number of OTUs and :math:`N` is the total number of
+    individuals in the sample.
 
     Assumes log accumulation.
 
@@ -491,27 +559,29 @@ def margalef(counts):
        76-77.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return (observed_otus(counts) - 1) / np.log(counts.sum())
 
 
 @experimental(as_of="0.4.0")
 def mcintosh_d(counts):
-    """Calculate McIntosh dominance index D, which is defined as:
+    r"""Calculate McIntosh dominance index D.
+
+    McIntosh dominance index D is defined as:
 
     .. math::
 
-       D = \\frac{N - U}{N - \\sqrt{N}}
+       D = \frac{N - U}{N - \sqrt{N}}
 
-    where :math:`N` is the total number of individuals (sum of abundances for
-    all OTUs) and :math:`U` is given as:
+    where :math:`N` is the total number of individuals in the sample and
+    :math:`U` is defined as:
 
     .. math::
 
-        U = \\sqrt{\\sum{{n_i}^2}}
+       U = \sqrt{\sum{{n_i}^2}}
 
-    where :math:`n_i` is the sum of abundances for all OTUs in the
-    :math:`i_{th}` species.
+    where :math:`n_i` is the number of individuals in the :math:`i^{\text{th}}`
+    OTU.
 
     Parameters
     ----------
@@ -539,7 +609,7 @@ def mcintosh_d(counts):
     .. [2] http://www.pisces-conservation.com/sdrhelp/index.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     u = np.sqrt((counts * counts).sum())
     n = counts.sum()
     return (n - u) / (n - np.sqrt(n))
@@ -547,7 +617,17 @@ def mcintosh_d(counts):
 
 @experimental(as_of="0.4.0")
 def mcintosh_e(counts):
-    """Calculate McIntosh's evenness measure E.
+    r"""Calculate McIntosh's evenness measure E.
+
+    McIntosh evenness measure E is defined as:
+
+    .. math::
+
+       E = \frac{\sqrt{\sum{n_i^2}}}{\sqrt{((N-S+1)^2 + S -1}}
+
+    where :math:`n_i` is the number of individuals in the :math:`i^{\text{th}}`
+    OTU, :math:`N` is the total number of individuals, and :math:`S` is the
+    number of OTUs in the sample.
 
     Parameters
     ----------
@@ -565,15 +645,16 @@ def mcintosh_e(counts):
 
     Notes
     -----
-    The implementation here is based on the description given in [1]_, *NOT*
+    The implementation here is based on the description given in [1]_, **NOT**
     the one in the SDR-IV online manual, which is wrong.
 
     References
     ----------
-    .. [1] Heip & Engels 1974 p 560.
+    .. [1] Heip & Engels (1974) Comparing Species Diversity and Evenness
+       Indices. p 560.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     numerator = np.sqrt((counts * counts).sum())
     n = counts.sum()
     s = observed_otus(counts)
@@ -583,7 +664,16 @@ def mcintosh_e(counts):
 
 @experimental(as_of="0.4.0")
 def menhinick(counts):
-    """Calculate Menhinick's richness index.
+    r"""Calculate Menhinick's richness index.
+
+    Menhinick's richness index is defined as:
+
+    .. math::
+
+       D_{Mn} = \frac{S}{\sqrt{N}}
+
+    where :math:`S` is the number of OTUs and :math:`N` is the total number of
+    individuals in the sample.
 
     Assumes square-root accumulation.
 
@@ -607,19 +697,19 @@ def menhinick(counts):
        76-77.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return observed_otus(counts) / np.sqrt(counts.sum())
 
 
 @experimental(as_of="0.4.0")
 def michaelis_menten_fit(counts, num_repeats=1, params_guess=None):
-    """Calculate Michaelis-Menten fit to rarefaction curve of observed OTUs.
+    r"""Calculate Michaelis-Menten fit to rarefaction curve of observed OTUs.
 
-    The Michaelis-Menten equation is defined as
+    The Michaelis-Menten equation is defined as:
 
     .. math::
 
-       S=\\frac{nS_{max}}{n+B}
+       S=\frac{nS_{max}}{n+B}
 
     where :math:`n` is the number of individuals and :math:`S` is the number of
     OTUs. This function estimates the :math:`S_{max}` parameter.
@@ -665,7 +755,7 @@ def michaelis_menten_fit(counts, num_repeats=1, params_guess=None):
        Michaelis-Menten equation. Biometrics 43, 793-803.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
 
     n_indiv = counts.sum()
     if params_guess is None:
@@ -705,7 +795,7 @@ def observed_otus(counts):
         Distinct OTU count.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return (counts != 0).sum()
 
 
@@ -735,19 +825,19 @@ def osd(counts):
     on these three measures.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return observed_otus(counts), singles(counts), doubles(counts)
 
 
 @experimental(as_of="0.4.0")
 def robbins(counts):
-    """Calculate Robbins' estimator for the probability of unobserved outcomes.
+    r"""Calculate Robbins' estimator for the probability of unobserved outcomes.
 
-    Robbins' estimator is defined as
+    Robbins' estimator is defined as:
 
     .. math::
 
-       \\frac{F_1}{n+1}
+       \frac{F_1}{n+1}
 
     where :math:`F_1` is the number of singleton OTUs.
 
@@ -771,13 +861,22 @@ def robbins(counts):
     .. [1] Robbins, H. E (1968). Ann. of Stats. Vol 36, pp. 256-257.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return singles(counts) / counts.sum()
 
 
 @experimental(as_of="0.4.0")
 def shannon(counts, base=2):
-    """Calculate Shannon entropy of counts (H), default in bits.
+    r"""Calculate Shannon entropy of counts, default in bits.
+
+    Shannon-Wiener diversity index is defined as:
+
+    .. math::
+
+       H = -\sum_{i=1}^s\left(p_i\log_2 p_i\right)
+
+    where :math:`s` is the number of OTUs and :math:`p_i` is the proportion of
+    the community represented by OTU :math:`i`.
 
     Parameters
     ----------
@@ -794,7 +893,7 @@ def shannon(counts, base=2):
     Notes
     -----
     The implementation here is based on the description given in the SDR-IV
-    online manual [1]_, except that the default logarithm base used here is 2
+    online manual [1]_ except that the default logarithm base used here is 2
     instead of :math:`e`.
 
     References
@@ -802,7 +901,7 @@ def shannon(counts, base=2):
     .. [1] http://www.pisces-conservation.com/sdrhelp/index.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     freqs = counts / counts.sum()
     nonzero_freqs = freqs[freqs.nonzero()]
     return -(nonzero_freqs * np.log(nonzero_freqs)).sum() / np.log(base)
@@ -810,9 +909,16 @@ def shannon(counts, base=2):
 
 @experimental(as_of="0.4.0")
 def simpson(counts):
-    """Calculate Simpson's index.
+    r"""Calculate Simpson's index.
 
-    Simpson's index is defined as 1 - dominance.
+    Simpson's index is defined as ``1 - dominance``:
+
+    .. math::
+
+       1 - \sum{p_i^2}
+
+    where :math:`p_i` is the proportion of the community represented by OTU
+    :math:`i`.
 
     Parameters
     ----------
@@ -839,19 +945,19 @@ def simpson(counts):
     .. [2] http://www.pisces-conservation.com/sdrhelp/index.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return 1 - dominance(counts)
 
 
 @experimental(as_of="0.4.0")
 def simpson_e(counts):
-    """Calculate Simpson's evenness measure E.
+    r"""Calculate Simpson's evenness measure E.
 
     Simpson's E is defined as
 
     .. math::
 
-       E=\\frac{1 / D}{S_{obs}}
+       E=\frac{1 / D}{S_{obs}}
 
     where :math:`D` is dominance and :math:`S_{obs}` is the number of observed
     OTUs.
@@ -881,7 +987,7 @@ def simpson_e(counts):
     .. [1] http://www.tiem.utk.edu/~gross/bioed/bealsmodules/simpsonDI.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return enspie(counts) / observed_otus(counts)
 
 
@@ -900,13 +1006,26 @@ def singles(counts):
         Singleton count.
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     return (counts == 1).sum()
 
 
 @experimental(as_of="0.4.0")
 def strong(counts):
-    """Calculate Strong's dominance index (Dw).
+    r"""Calculate Strong's dominance index.
+
+    Strong's dominance index is defined as:
+
+    .. math::
+
+       D_w = max_i[(\frac{b_i}{N})-\frac{i}{S}]
+
+    where :math:`b_i` is the sequential cumulative totaling of the
+    :math:`i^{\text{th}}` OTU abundance values ranked from largest to smallest,
+    :math:`N` is the total number of individuals in the sample, and
+    :math:`S` is the number of OTUs in the sample. The expression in brackets
+    is computed for all OTUs, and :math:`max_i` denotes the maximum value in
+    brackets for any OTU.
 
     Parameters
     ----------
@@ -930,7 +1049,7 @@ def strong(counts):
     .. [2] http://www.pisces-conservation.com/sdrhelp/index.html
 
     """
-    counts = _validate(counts)
+    counts = _validate_counts_vector(counts)
     n = counts.sum()
     s = observed_otus(counts)
     i = np.arange(1, len(counts) + 1)