+++ skbio/diversity/tests/__init__.py
@@ -6,14 +6,4 @@
 # The full license is in the file COPYING.txt, distributed with this software.
 # ----------------------------------------------------------------------------
 
-ifeq ($(WITH_COVERAGE), TRUE)
-	TEST_COMMAND = coverage run -m skbio.test
-else
-	TEST_COMMAND = python -m skbio.test
-endif
-
-test:
-	$(TEST_COMMAND)
-	flake8 skbio setup.py checklist.py
-	./checklist.py
-	check-manifest
+from __future__ import absolute_import, division, print_function