+++ skbio/tree/_tree.py
@@ -8,6 +8,7 @@
 
 from __future__ import absolute_import, division, print_function
 
+import warnings
 from operator import or_
 from copy import deepcopy
 from itertools import combinations
@@ -23,6 +24,7 @@ from skbio._base import SkbioObject
 from skbio.stats.distance import DistanceMatrix
 from ._exception import (NoLengthError, DuplicateNodeError, NoParentError,
                          MissingNodeError, TreeError)
+from skbio.util import RepresentationWarning
 from skbio.util._decorator import experimental
 
 
@@ -438,7 +440,7 @@ class TreeNode(SkbioObject):
         ids = set(names)
 
         if not ids.issubset(all_tips):
-            raise ValueError("ids are not a subset of the tree!")
+            raise ValueError("ids are not a subset of the tree.")
 
         while len(list(tcopy.tips())) != len(ids):
             for n in list(tcopy.tips()):
@@ -649,6 +651,46 @@ class TreeNode(SkbioObject):
             return len(list(self.tips()))
         else:
             return len(list(self.traverse(include_self=True)))
+
+    @experimental(as_of="0.4.0")
+    def observed_node_counts(self, tip_counts):
+        """Returns counts of node observations from counts of tip observations
+
+        Parameters
+        ----------
+        tip_counts : dict of ints
+            Counts of observations of tips. Keys correspond to tip names in
+            ``self``, and counts are unsigned ints.
+
+        Returns
+        -------
+        dict
+            Counts of observations of nodes. Keys correspond to node names
+            (internal nodes or tips), and counts are unsigned ints.
+
+        Raises
+        ------
+        ValueError
+            If a count less than one is observed.
+        MissingNodeError
+            If a count is provided for a tip not in the tree, or for an
+            internal node.
+
+        """
+        result = defaultdict(int)
+        for tip_name, count in tip_counts.items():
+            if count < 1:
+                raise ValueError("All tip counts must be greater than zero.")
+            else:
+                t = self.find(tip_name)
+                if not t.is_tip():
+                    raise MissingNodeError(
+                        "Counts can only be for tips in the tree. %s is an "
+                        "internal node." % t.name)
+                result[t] += count
+                for internal_node in t.ancestors():
+                    result[internal_node] += count
+        return result
 
     @experimental(as_of="0.4.0")
     def subtree(self, tip_list=None):
@@ -1370,7 +1412,7 @@ class TreeNode(SkbioObject):
                 if node.is_tip():
                     if name in tip_cache:
                         raise DuplicateNodeError("Tip with name '%s' already "
-                                                 "exists!" % name)
+                                                 "exists." % name)
 
                     tip_cache[name] = node
                 else:
@@ -1746,7 +1788,7 @@ class TreeNode(SkbioObject):
         tips = [self.find(name) for name in tipnames]
 
         if len(tips) == 0:
-            raise ValueError("No tips found!")
+            raise ValueError("No tips found.")
 
         nodes_to_scrub = []
 
@@ -2205,7 +2247,7 @@ class TreeNode(SkbioObject):
                 raise NoParentError("Provided ancestor is not in the path")
 
             if curr.length is None:
-                raise NoLengthError("No length on node %s found!" %
+                raise NoLengthError("No length on node %s found." %
                                     curr.name or "unnamed")
 
             accum += curr.length
@@ -2369,13 +2411,16 @@ class TreeNode(SkbioObject):
         ------
         ValueError
             If any of the specified `endpoints` are not tips
-        NoLengthError
-            If a node without length is encountered
 
         See Also
         --------
         distance
         compare_tip_distances
+
+        Notes
+        -----
+        If a node does not have an associated length, 0.0 will be used and a
+        ``RepresentationWarning`` will be raised.
 
         Examples
         --------
@@ -2436,11 +2481,14 @@ class TreeNode(SkbioObject):
             # can possibly use np.zeros
             starts, stops = [], []  # to calc ._start and ._stop for curr node
             for child in node.children:
-                if child.length is None:
-                    raise NoLengthError("Node with name '%s' doesn't have a "
-                                        "length." % child.name)
-
-                distances[child.__start:child.__stop] += child.length
+                length = child.length
+                if length is None:
+                    warnings.warn(
+                        "`TreeNode.tip_tip_distances`: Node with name %r does "
+                        "not have an associated length, so a length of 0.0 "
+                        "will be used." % child.name, RepresentationWarning)
+                    length = 0.0
+                distances[child.__start:child.__stop] += length
 
                 starts.append(child.__start)
                 stops.append(child.__stop)
@@ -2837,7 +2885,7 @@ class TreeNode(SkbioObject):
                 return a + b
 
         else:
-            raise TypeError("Only list, set and frozenset are supported!")
+            raise TypeError("Only list, set and frozenset are supported.")
 
         for node in self.postorder(include_self=True):
             node._registered_caches.add(cache_attrname)